+++ NEWS.md
@@ -1,3 +1,20 @@
+# Rsagacmd 0.2
+
+New features:
+
+- Support for SpatVector and SpatVectorProxy proxy objects for vector spatial
+datasets.
+
+- tidy methods to summarize information about the SAGA-GIS libraries and tools
+within any library and return these as a tibble.
+
+- Function 'saga_docs' to browse the online documentation for each SAGA-GIS
+tool.
+
+Changes:
+
+ - Bug fix for using the `terra` package as a raster backend.
+
 # Rsagacmd 0.1.2
 
 New features: